// src/shared/types.ts
export type RectSize = {
  width: DOMRect['width'];
  height: DOMRect['height'];
};

export type OnBoardGeometry = {
  boardX: number;
  boardY: number;
  width: DOMRect['width'];
  height: DOMRect['height'];
};

export type OnClientGeometry = {
  clientX: number;
  clientY: number;
  width: DOMRect['width'];
  height: DOMRect['height'];
};